Update games.heist.json
Changed
"Some loss of heist team is nothing compared to remaining crew have in their pockets."
to
"Some loss of heist team is nothing compared to what remaining crew have in their pockets."
The "what" was there earlier, but was removed when other things were fixed. It needs to stay.